pyproject.toml
[project]
name = "hungrier-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-ext-autoreload>=0.1.2",
    "discord-py>=2.5.2",
]
